+++ web/src/components/organisms/Project/Accessibility/Accessibility/hooks.ts
@@ -3,8 +3,8 @@ import { useCallback, useMemo, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 
 import Notification from "@reearth-cms/components/atoms/Notification";
-import { FormType } from "@reearth-cms/components/molecules/Accessibility/types";
 import { Model } from "@reearth-cms/components/molecules/Model/types";
+import { FormType } from "@reearth-cms/components/molecules/PublicAPI/types";
 import { fromGraphQLModel } from "@reearth-cms/components/organisms/DataConverters/model";
 import { Model as GQLModel } from "@reearth-cms/gql/__generated__/graphql.generated";
 import { GetModelsDocument } from "@reearth-cms/gql/__generated__/model.generated";
